+++ packages/sphinx-test-reports/tests/doc_test/doc_test_file/conf.py
@@ -18,6 +18,9 @@
 import os
 import sys
 
+import sphinx_needs
+from packaging.version import Version
+
 sys.path.insert(0, os.path.abspath("../../sphinxcontrib"))
 
 # -- General configuration ------------------------------------------------
@@ -75,7 +78,14 @@ source_suffix = ".rst"
 # The master toctree document.
 master_doc = "index"
 
-needs_extra_options = ["asil", "uses_secure", "more_info"]
+if Version(sphinx_needs.__version__) >= Version("8.0.0"):
+    needs_fields = {
+        "asil": {"nullable": True},
+        "uses_secure": {"nullable": True},
+        "more_info": {"nullable": True},
+    }
+else:
+    needs_extra_options = ["asil", "uses_secure", "more_info"]
 tr_extra_options = ["more_info"]
 
 # General information about the project.