+++ packages/sphinx-test-reports/tests/doc_test/many_testsuites_doc/conf.py
@@ -18,6 +18,9 @@
 import os
 import sys
 
+import sphinx_needs
+from packaging.version import Version
+
 sys.path.insert(0, os.path.abspath("../../sphinxcontrib"))
 
 # -- General configuration ------------------------------------------------
@@ -37,7 +40,10 @@ tr_case_id_length = 10
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
-needs_extra_options = ["more_info"]
+if Version(sphinx_needs.__version__) >= Version("8.0.0"):
+    needs_fields = {"more_info": {"nullable": True}}
+else:
+    needs_extra_options = ["more_info"]
 tr_extra_options = ["more_info"]
 
 # The suffix(es) of source filenames.