+++ packages/sphinx-test-reports/docs/changelog.rst
@@ -107,7 +107,7 @@ introduces several maintenance improvements.
 -----
 :Released: 26.09.2022
 
-* Improvement: Supporting `Sphinx-Needs <https://www.sphinx-needs.com/>`__ ``>= 1.01`` only.
+* Improvement: Supporting `Sphinx-Needs <https://sphinx-needs.readthedocs.io/en/latest/>`__ ``>= 1.01`` only.
 * Improvement: Migrated nosetests to pytest.
 
 0.3.7